requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scipy>=1.7.0

# API and model dependencies
openai>=1.0.0
anthropic>=0.20.0
transformers>=4.30.0
torch>=2.0.0
tiktoken>=0.5.0

# Statistical analysis
scikit-learn>=1.0.0
statsmodels>=0.13.0

# Utilities
tqdm>=4.60.0
python-dotenv>=0.19.0
requests>=2.25.0
jsonlines>=3.0.0

# Jupyter support (optional, for notebooks)
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.4.0

# Data manipulation
openpyxl>=3.0.0  # For Excel file support
xlrd>=2.0.0      # For reading Excel files

# Visualization extras
plotly>=5.0.0    # For interactive plots
pillow>=8.0.0    # For image processing